Use type uint8 for boolean arrays.

+++ python/boomer/algorithm/_label_wise_measure.pxd
@@ -14,7 +14,7 @@ cdef class LabelWiseMeasure(DecomposableLoss):
 
     cdef readonly float64[::1, :] coverable_labels
 
-    cdef uint32[::1] minority_labels
+    cdef uint8[::1] minority_labels
 
     cdef uint8[::1, :] true_labels
